+++ frontend/src/scenes/data-warehouse/external/forms/PostgresSchemaForm.tsx
@@ -1,13 +1,13 @@
 import { LemonSwitch, LemonTable } from '@posthog/lemon-ui'
 import { useActions, useMountedLogic, useValues } from 'kea'
 
-import { sourceModalLogic } from '../sourceModalLogic'
+import { sourceWizardLogic } from '../../new/sourceWizardLogic'
 import { sourceFormLogic } from './sourceFormLogic'
 
 export default function PostgresSchemaForm(): JSX.Element {
     useMountedLogic(sourceFormLogic({ sourceType: 'Postgres' }))
-    const { selectSchema } = useActions(sourceModalLogic)
-    const { databaseSchema } = useValues(sourceModalLogic)
+    const { selectSchema } = useActions(sourceWizardLogic)
+    const { databaseSchema } = useValues(sourceWizardLogic)
 
     return (
         <div className="flex flex-col gap-2">
